requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
cloudinary==1.29.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.14
django-cors-headers==3.7.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-decouple==3.6
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.10
whitenoise==6.2.0
zope.interface==5.4.0
